+++ src/repo/mongoose/model/transaction.ts
@@ -2,60 +2,6 @@ import * as mongoose from 'mongoose';
 
 const writeConcern: mongoose.WriteConcern = { j: true, w: 'majority', wtimeout: 10000 };
 
-const objectSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const resultSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const agentSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const recipientSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const errorSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
-const potentialActionsSchema = new mongoose.Schema(
-    {},
-    {
-        id: false,
-        _id: false,
-        strict: false
-    }
-);
-
 /**
  * 取引スキーマ
  */
@@ -64,17 +10,17 @@ const schema = new mongoose.Schema(
         project: mongoose.SchemaTypes.Mixed,
         status: String,
         typeOf: String,
-        agent: agentSchema,
-        recipient: recipientSchema,
-        error: errorSchema,
-        result: resultSchema,
-        object: objectSchema,
+        agent: mongoose.SchemaTypes.Mixed,
+        recipient: mongoose.SchemaTypes.Mixed,
+        error: mongoose.SchemaTypes.Mixed,
+        result: mongoose.SchemaTypes.Mixed,
+        object: mongoose.SchemaTypes.Mixed,
         expires: Date,
         startDate: Date,
         endDate: Date,
         tasksExportedAt: Date,
         tasksExportationStatus: String,
-        potentialActions: potentialActionsSchema
+        potentialActions: mongoose.SchemaTypes.Mixed
     },
     {
         collection: 'transactions',
@@ -87,8 +33,18 @@ const schema = new mongoose.Schema(
             createdAt: 'createdAt',
             updatedAt: 'updatedAt'
         },
-        toJSON: { getters: true },
-        toObject: { getters: true }
+        toJSON: {
+            getters: false,
+            virtuals: false,
+            minimize: false,
+            versionKey: false
+        },
+        toObject: {
+            getters: false,
+            virtuals: true,
+            minimize: false,
+            versionKey: false
+        }
     }
 );