+++ src/repo/transaction.ts
@@ -318,7 +318,7 @@ export class MongoRepository {
             .toDate();
 
         // ステータスと期限を見て更新
-        await this.transactionModel.update(
+        await this.transactionModel.updateMany(
             {
                 status: factory.transactionStatusType.InProgress,
                 expires: { $lt: endDate }
@@ -326,8 +326,7 @@ export class MongoRepository {
             {
                 status: factory.transactionStatusType.Expired,
                 endDate: endDate
-            },
-            { multi: true }
+            }
         )
             .exec();
     }